indfor_uniforms - split up config into includes, add dummy config for tanker uniform
