+++ addons/indfor_uniforms/CfgWeapons.hpp
@@ -3,148 +3,7 @@ class CfgWeapons
 	class ItemCore;
 	class UniformItem;
 	class Uniform_Base;
-	class tmtm_u_commandoSweater_black: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Black)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_black_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_black";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_blue: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Blue)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_blue_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_blue";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_brown: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Brown)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_brown_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_brown";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_erdl: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (ERDL)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_olive_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_erdl";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_erdlArvn: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (ERDL ARVN)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_black_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_erdlArvn";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_erdlOcean: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (ERDL Ocean)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_navy_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_erdlOcean";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_ghost: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Ghost)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_navy_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_ghost";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_navy: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Navy)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_navy_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_navy";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
-	class tmtm_u_commandoSweater_red: Uniform_Base
-	{
-		scope=1;
-		scopeArsenal=2;
-		scopeCurator=0;
-		displayName="Commando Sweater (Red)";
-		picture="\tmtm\pack\indfor_uniforms\Data\UI\icon_u_commandoSweater_red_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_commandoSweater_red";
-			containerClass="Supply50";
-			mass=50;
-		};
-	};
+
+	#include "include/commandoSweater/CfgWeapons.hpp"
+	#include "include/tankOveralls/CfgWeapons.hpp"
 };
